Add files via upload
SP51_SRPY_Attack - was the code that was used to fly on October 6, 2017.  SP stands for SensorPod, which represents all of the sensors that are interfaced to the Raspberry Pi.  The .npy files are convolution kernels that are used in the vision code.  SRPY stands for the fact that this code controls a vertical camera gimbal servo, the vehicle's Roll, Pitch, and Yaw axes.  The 'Attack' stands for the code that executes the terminal attack maneuver and counter-measure release.